manifest: move Group code into a file of its own
